Optimisation round1

+++ src/cray/crconsts.h
@@ -13,6 +13,6 @@ static const double PLAYER_ARROW_SIZE = 25.0;
 static const double PLAYER_BASE_SIZE = 6.0;
 static const double TRANS_SPEED = 2.0;
 static const double ROT_SPEED = 2.0;
-static const double WALL_HEIGHT = 4000.0;
+static const int WALL_HEIGHT = 4000.0;
 
 #endif // !_CR_CONSTS_H_
